+++ microservice/microservice/dbmicroservice/MicroserviceBootstrapper.cs
@@ -61,59 +61,12 @@ namespace Beamable.Server
 	    public static ContentService ContentService;
 	    public static List<BeamableMicroService> Instances = new List<BeamableMicroService>();
 
-	    public static bool TryParseLogLevel(string logLevel, out LogEventLevel serilogLevel)
-	    {
-		    if (logLevel == null)
-		    {
-			    serilogLevel= LogEventLevel.Debug;
-			    return false;
-		    }
-
-		    switch (logLevel.ToLowerInvariant())
-		    {
-			    
-			    case "f":
-			    case "fatal":
-				    serilogLevel = LogEventLevel.Fatal;
-				    return true;
-			    case "e":
-			    case "err":
-			    case "error":
-				    serilogLevel = LogEventLevel.Error;
-				    return true;
-			    case "v":
-			    case "verbose":
-				    serilogLevel = LogEventLevel.Verbose;
-				    return true;
-			    case "d":
-			    case "dbug":
-			    case "dbg":
-
-			    case "debug":
-				    serilogLevel = LogEventLevel.Debug;
-				    return true;
-			    case "w":
-			    case "warn":
-			    case "warning":
-				    serilogLevel = LogEventLevel.Warning;
-				    return true;
-			    case "i":
-			    case "information":
-			    case "info":
-				    serilogLevel = LogEventLevel.Information;
-				    return true;
-			    default:
-				    serilogLevel = LogEventLevel.Debug;
-				    return false;
-		    }
-	    }
-	    
-        private static void ConfigureLogging(IMicroserviceArgs args)
+	    private static DebugLogSink ConfigureLogging(IMicroserviceArgs args, MicroserviceAttribute attr)
         {
             var logLevel = args.LogLevel;
 			var disableLogTruncate = (Environment.GetEnvironmentVariable("DISABLE_LOG_TRUNCATE")?.ToLowerInvariant() ?? "") == "true";
 
-			TryParseLogLevel(logLevel, out var envLogLevel);
+			LogUtil.TryParseLogLevel(logLevel, out var envLogLevel);
 
             var inDocker = Environment.GetEnvironmentVariable("DOTNET_RUNNING_IN_CONTAINER") == "true";
             
@@ -147,6 +100,14 @@ namespace Beamable.Server
             }
 
             var logger = logConfig;
+
+            DebugLogSink debugLogSink = null;
+            if (!inDocker)
+            {
+	            debugLogSink = new DebugLogSink(new MicroserviceLogFormatter());
+	            logConfig.WriteTo.Sink(debugLogSink);
+            }
+            
             switch (args.LogOutputType)
             {
 	            case LogOutputType.DEFAULT when !inDocker:
@@ -177,6 +138,8 @@ namespace Beamable.Server
             BeamableLogProvider.Provider = new BeamableSerilogProvider();
             Debug.Instance = new MicroserviceDebug();
             BeamableSerilogProvider.LogContext.Value = Log.Logger;
+
+            return debugLogSink;
         }
 
         public static void ConfigureUnhandledError()
@@ -451,8 +414,9 @@ namespace Beamable.Server
         public static async Task Start<TMicroService>() where TMicroService : Microservice
         {
 	        var attribute = typeof(TMicroService).GetCustomAttribute<MicroserviceAttribute>();
-	        var envArgs = new EnvironmentArgs();
-	        ConfigureLogging(envArgs);
+
+	        var envArgs = new EnviornmentArgs();
+	        var pipeSink = ConfigureLogging(envArgs, attribute);
 	        ConfigureUncaughtExceptions();
 	        ConfigureUnhandledError();
 	        ConfigureDiscovery(envArgs, attribute);
@@ -487,7 +451,7 @@ namespace Beamable.Server
 				
 	            if (isFirstInstance)
 	            {
-		            var localDebug = new ContainerDiagnosticService(instanceArgs, beamableService);
+		            var localDebug = new ContainerDiagnosticService(instanceArgs, beamableService, pipeSink);
 		            var runningDebugTask = localDebug.Run();
 	            }