+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -5,13 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 ### Added
 - MongoDbExtensions class that supports mongo indexes creation
 - `ICollectionElement` interface and `MongoIndexAttribute` to support automatic index creation during microservice startup
+- Set of CRUD methods for `IMongoCollection` and `IStorageObjectConnectionProvider`
 
-## [1.15.2]
+## [1.16.0]
 
 no changes